refactor: mv `IOActions` up from `src/Model/` into `src/`
The new location makes it easier for other modules to access types like
`IODirection`, and also better reflects their importance to the package.
