+++ src/IOActions/aliasMapping.ts
@@ -1,7 +1,7 @@
 import { hasKey } from "@nerdware/ts-type-safety-utils";
-import { ItemInputError } from "../../utils/index.js";
-import type { IOActions, IOAction } from "./types.js";
-import type { BaseItem } from "../../types/index.js";
+import { ItemInputError } from "../utils/index.js";
+import type { IOAction } from "./types/index.js";
+import type { BaseItem } from "../types/index.js";
 
 /**
  * This `IOAction` swaps attribute-names with their corresponding aliases:
@@ -11,7 +11,7 @@ import type { BaseItem } from "../../types/index.js";
  * @throws If an attribute is not defined in the schema, and `allowUnknownAttributes` is not `true`.
  */
 export const aliasMapping: IOAction = function (
-  this: IOActions,
+  this,
   item,
   { aliasesMap, modelName, schema, schemaOptions: { allowUnknownAttributes } }
 ) {
