+++ src/IOActions/types/IODirection.ts
@@ -1,0 +1,5 @@
+/**
+ * Labels corresponding to the request-response cycle which indicate the
+ * "direction" data is flowing — either to or from the database.
+ */
+export type IODirection = "toDB" | "fromDB";
